Rename kmm.viewmodel to kmp.observableviewmodel
