--- kmp-observableviewmodel-core/src/commonMain/kotlin/com/rickclephas/kmm/viewmodel/InternalKMMViewModelApi.kt
@@ -1,9 +1,0 @@
-package com.rickclephas.kmm.viewmodel
-
-@RequiresOptIn(
-    level = RequiresOptIn.Level.ERROR,
-    message = "This is an internal KMM-ViewModel API that shouldn't be used outside KMM-ViewModel!"
-)
-@Retention(value = AnnotationRetention.BINARY)
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY)
-public annotation class InternalKMMViewModelApi
