--- kmp-observableviewmodel-core/src/commonMain/kotlin/com/rickclephas/kmm/viewmodel/ViewModelScope.kt
@@ -1,19 +1,0 @@
-package com.rickclephas.kmm.viewmodel
-
-import kotlinx.coroutines.CoroutineScope
-
-/**
- * Holds the [CoroutineScope] of a [KMMViewModel].
- * @see coroutineScope
- */
-public expect interface ViewModelScope
-
-/**
- * Creates a new [ViewModelScope] for the provided [coroutineScope].
- */
-internal expect fun ViewModelScope(coroutineScope: CoroutineScope): ViewModelScope
-
-/**
- * Gets the [CoroutineScope] associated with the [KMMViewModel] of `this` [ViewModelScope].
- */
-public expect val ViewModelScope.coroutineScope: CoroutineScope
