+++ kmp-observableviewmodel-core/src/commonMain/kotlin/com/rickclephas/kmp/observableviewmodel/InternalKMMViewModelApi.kt
@@ -1,0 +1,9 @@
+package com.rickclephas.kmp.observableviewmodel
+
+@RequiresOptIn(
+    level = RequiresOptIn.Level.ERROR,
+    message = "This is an internal KMM-ViewModel API that shouldn't be used outside KMM-ViewModel!"
+)
+@Retention(value = AnnotationRetention.BINARY)
+@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY)
+public annotation class InternalKMMViewModelApi
