+++ kmp-observableviewmodel-core/src/nativeInterop/cinterop/KMMViewModelCoreObjC.def
@@ -1,3 +1,3 @@
 language = Objective-C
-package = com.rickclephas.kmm.viewmodel.objc
+package = com.rickclephas.kmp.observableviewmodel.objc
 headers = KMMVMViewModelScope.h
