--- sample/androidApp/src/main/java/com/rickclephas/kmm/viewmodel/sample/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.rickclephas.kmm.viewmodel.sample
-
-import androidx.fragment.app.FragmentActivity
-
-class MainActivity : FragmentActivity(R.layout.activity_main)
