+++ sample/androidApp/src/main/java/com/rickclephas/kmp/observableviewmodel/sample/ComposeFragment.kt
@@ -1,0 +1,36 @@
+package com.rickclephas.kmp.observableviewmodel.sample
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.fragment.app.Fragment
+import com.rickclephas.kmp.observableviewmodel.sample.shared.Greeting
+
+class ComposeFragment: Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View = ComposeView(requireContext()).apply {
+        setContent {
+            Greeting(Greeting().greeting())
+        }
+    }
+}
+
+@Composable
+fun Greeting(text: String) {
+    Text(text = text)
+}
+
+@Preview
+@Composable
+fun DefaultPreview() {
+    Greeting("Hello, Android!")
+}
