+++ sample/androidApp/src/main/java/com/rickclephas/kmp/observableviewmodel/sample/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.rickclephas.kmp.observableviewmodel.sample
+
+import androidx.fragment.app.FragmentActivity
+
+class MainActivity : FragmentActivity(R.layout.activity_main)
