--- sample/shared/src/androidMain/kotlin/com/rickclephas/kmm/viewmodel/sample/shared/Platform.kt
@@ -1,7 +1,0 @@
-package com.rickclephas.kmm.viewmodel.sample.shared
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
