+++ sample/shared/src/androidMain/kotlin/com/rickclephas/kmp/observableviewmodel/sample/shared/Platform.kt
@@ -1,0 +1,7 @@
+package com.rickclephas.kmp.observableviewmodel.sample.shared
+
+class AndroidPlatform : Platform {
+    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform(): Platform = AndroidPlatform()
