--- sample/shared/src/commonMain/kotlin/com/rickclephas/kmm/viewmodel/sample/shared/Greeting.kt
@@ -1,9 +1,0 @@
-package com.rickclephas.kmm.viewmodel.sample.shared
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greeting(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
