--- sample/shared/src/commonMain/kotlin/com/rickclephas/kmm/viewmodel/sample/shared/Platform.kt
@@ -1,7 +1,0 @@
-package com.rickclephas.kmm.viewmodel.sample.shared
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
