+++ sample/shared/src/commonMain/kotlin/com/rickclephas/kmp/observableviewmodel/sample/shared/Greeting.kt
@@ -1,0 +1,9 @@
+package com.rickclephas.kmp.observableviewmodel.sample.shared
+
+class Greeting {
+    private val platform: Platform = getPlatform()
+
+    fun greeting(): String {
+        return "Hello, ${platform.name}!"
+    }
+}
