+++ sample/shared/src/commonMain/kotlin/com/rickclephas/kmp/observableviewmodel/sample/shared/Platform.kt
@@ -1,0 +1,7 @@
+package com.rickclephas.kmp.observableviewmodel.sample.shared
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
