--- sample/shared/src/iosMain/kotlin/com/rickclephas/kmm/viewmodel/sample/shared/Platform.kt
@@ -1,9 +1,0 @@
-package com.rickclephas.kmm.viewmodel.sample.shared
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
