+++ sample/shared/src/iosMain/kotlin/com/rickclephas/kmp/observableviewmodel/sample/shared/Platform.kt
@@ -1,0 +1,9 @@
+package com.rickclephas.kmp.observableviewmodel.sample.shared
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
